재고관리 (새로운 제품 발주완료 시 products 테이블에 등록, 기존 상품 추가 발주 시 storage_qty 업데이트) 발주관리 (발주상태 업데이트 추가, 발주 처리 기능 추가) 불량재고 (발주 처리 시 불량재고가 있을 경우 불량재고 페이지에 업데이트 됨)

+++ teamcostco/src/main/java/com/ezentwix/teamcostco/dto/product/ProductDefectiveDTO.java
@@ -8,7 +8,7 @@ import lombok.Data;
 public class ProductDefectiveDTO {
 
     private Integer defective_id;
-    private Integer product_code;
+    private Long product_code;
     private String product_name;
     private Integer defective_qty;
     private String reason;
